requirements.txt
folium
pandas
requests
Django
dj-database-url
whitenoise
psycopg2-binary
gunicorn
python-decouple
praw
python-dotenv
django-cors-headers